Load Types (#116)
* add result status enum

* move load result out into its own class

* load changes

* add search result value + logic for it

* document load types

* removed isPlaylist field from docs

* add checks for unknown value

* remove isPlaylist field

* clarify version changes

* add references to pr's
