+++ LavalinkServer/src/main/java/lavalink/server/player/AudioLoader.java
@@ -31,7 +31,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 public class AudioLoader implements AudioLoadResultHandler {
@@ -40,9 +39,9 @@ public class AudioLoader implements AudioLoadResultHandler {
     private final AudioPlayerManager audioPlayerManager;
 
     private List<AudioTrack> loadedItems;
-    private boolean isPlaylist = false;
     private String playlistName = null;
     private Integer selectedTrack = null;
+    private ResultStatus status = ResultStatus.UNKNOWN;
     private boolean used = false;
 
     public AudioLoader(AudioPlayerManager audioPlayerManager) {
@@ -61,13 +60,16 @@ public class AudioLoader implements AudioLoadResultHandler {
             this.wait();
         }
 
-        return new LoadResult(loadedItems, isPlaylist, playlistName, selectedTrack);
+        if (status == ResultStatus.UNKNOWN)
+            throw new IllegalStateException("Load Type == UNKNOWN (shouldn't happen!)");
+        return new LoadResult(loadedItems, playlistName, status, selectedTrack);
     }
 
     @Override
     public void trackLoaded(AudioTrack audioTrack) {
         loadedItems = new ArrayList<>();
         loadedItems.add(audioTrack);
+        status = ResultStatus.TRACK_LOADED;
         log.info("Loaded track " + audioTrack.getInfo().title);
         synchronized (this) {
             this.notify();
@@ -77,12 +79,12 @@ public class AudioLoader implements AudioLoadResultHandler {
     @Override
     public void playlistLoaded(AudioPlaylist audioPlaylist) {
         if (!audioPlaylist.isSearchResult()) {
-            isPlaylist = true;
             playlistName = audioPlaylist.getName();
             selectedTrack = audioPlaylist.getTracks().indexOf(audioPlaylist.getSelectedTrack());
         }
 
         log.info("Loaded playlist " + audioPlaylist.getName());
+        status = audioPlaylist.isSearchResult() ? ResultStatus.SEARCH_RESULT : ResultStatus.PLAYLIST_LOADED;
         loadedItems = audioPlaylist.getTracks();
         synchronized (this) {
             this.notify();
@@ -92,6 +94,7 @@ public class AudioLoader implements AudioLoadResultHandler {
     @Override
     public void noMatches() {
         log.info("No matches found");
+        status = ResultStatus.NO_MATCHES;
         loadedItems = new ArrayList<>();
         synchronized (this) {
             this.notify();
@@ -101,24 +104,11 @@ public class AudioLoader implements AudioLoadResultHandler {
     @Override
     public void loadFailed(FriendlyException e) {
         log.error("Load failed", e);
+        status = ResultStatus.LOAD_FAILED;
         loadedItems = new ArrayList<>();
         synchronized (this) {
             this.notify();
         }
     }
 
-}
-
-class LoadResult {
-    public List<AudioTrack> tracks;
-    public boolean isPlaylist;
-    public String playlistName;
-    public Integer selectedTrack;
-
-    public LoadResult(List<AudioTrack> tracks, boolean isPlaylist, String playlistName, Integer selectedTrack) {
-        this.tracks = Collections.unmodifiableList(tracks);
-        this.isPlaylist = isPlaylist;
-        this.playlistName = playlistName;
-        this.selectedTrack = selectedTrack;
-    }
 }
