+++ LavalinkServer/src/main/java/lavalink/server/player/ResultStatus.java
@@ -1,0 +1,10 @@
+package lavalink.server.player;
+
+public enum ResultStatus {
+    TRACK_LOADED,
+    PLAYLIST_LOADED,
+    SEARCH_RESULT,
+    NO_MATCHES,
+    LOAD_FAILED,
+    UNKNOWN
+}
